% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ars.avg.plot.R
\name{ars.avg.plot}
\alias{ars.avg.plot}
\title{All features averages plot}
\usage{

  ars.avg.plot(indir = c("/home/suvar/Projects/007_ChIPseq_HU-SR/output/ars_dat/quantiles/Et",
  "/home/suvar/Projects/007_ChIPseq_HU-SR/output/ars_dat/quantiles/IAA"),
  expnames = c("HU-Et", "HU-IAA"), timepoints = c("2h", "1.5h", "0",
  "10", "20", "30", "40", "50", "60", "70", "80", "90"),
  splitnames = c("Q1", "Q2", "Q3", "Q4"),
  outdir = "/home/suvar/Desktop/", setmink = TRUE,
  title = "All features averages plot", xlab = "Time",
  ylab = "Signal intensity", filename = "ars_avg_plot")
}
\arguments{
\item{indir}{Input directory containing bedgraph files.}

\item{outdir}{Output directory.}

\item{title}{Main title of the plot}

\item{filename}{Name of the output plot file}

\item{fdir}{Directory containing different features in files.}

\item{chromsizes}{Text file containing sizes of all chromosomes in columns (chromosome ~ size).}

\item{notes}{Optional notes, which is displayed in the bottom of the plot.}

\item{exclude_seq}{Optional argument to exclude chromosomes.}
}
\description{
All features averages plot
}
