% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avg.timeseries.plot.R
\name{avg.timeseries.plot}
\alias{avg.timeseries.plot}
\title{Calculate matrix of averages from meandat files and plot in time series}
\usage{
avg.timeseries.plot(indir, expnames, timepoints, splitnames, outdir,
  title = "Average signal in time series plot", xlab = "Time",
  ylab = "Signal intensity", filename = "avg_timeseries_plot",
  frommin = 0, savemx = FALSE)
}
\arguments{
\item{indir}{Input director-y/-ies containing meandat files (xxx.mean.dat.gz). Files within one directory may be split by groups/quantiles etc.}

\item{expnames}{Names of experiments using for plot (must be the same as number of entry directories in indir).}

\item{timepoints}{Names of timepoints for plot (must be in the same order as in meandat files).}

\item{splitnames}{Names of split meandat files (e.g by groups or quantiles of replication time or expression levels).}

\item{outdir}{Output directory.}

\item{title}{Main title of the plot}

\item{filename}{Name of the output plot file}

\item{frommin}{How many bp skeep from minimal value. Useful if necessary to exclude a region wihin nucleosome free regions (NFR), e.g. near TSS or ARS.}

\item{savemx}{Save data matrix as csv for plots}
}
\description{
Calculate matrix of averages from meandat files and plot in time series
}
\author{
Mark Boltengagen \email{m.boltengagen@gmail.com}
}
