% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bamcov.plot.R
\name{bamcov.plot}
\alias{bamcov.plot}
\title{Bam Coverage Statter Plot}
\usage{
bamcov.plot(indir, outdir)
}
\arguments{
\item{indir}{Directory containing bam coverage binned bedgraph files}

\item{outdir}{Output directory}
}
\description{
Bam Coverage Statter Plot
}
