% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdg.binned.average.r
\name{bdg.binned.average}
\alias{bdg.binned.average}
\title{Binned average from bedgraph file}
\usage{
bdg.binned.average(indir, outdir, binsize = 100)
}
\arguments{
\item{indir}{Directory containing bedgraph files}

\item{outdir}{Output directory}

\item{binsize}{Length of one bin (default - 100 bp)}
}
\description{
Binned average from bedgraph file
}
\details{
Calculate binned average from bedgraph files
}
