% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdg.norm.R
\name{bdg.norm}
\alias{bdg.norm}
\title{Normalize bed graph files}
\usage{
bdg.norm(indir, outdir, normk = NULL)
}
\arguments{
\item{indir}{Directory containing bedgraph files with peaks}

\item{outdir}{Output directory to save bedgraph files after normalization}

\item{normk}{Vector of lengths of files containing multipliers for normalization}
}
\description{
Normalize bed graph files
}
\details{
Normalize bed graph files
}
