% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.mean.R
\name{check.coord}
\alias{check.coord}
\title{Check coordinates in dat files.}
\usage{
check.coord(datdir)
}
\arguments{
\item{dat}{Directory containing dat files.}
}
\value{
List of integers: \item{upstream}{description upstream}, \item{downstream}{description downstream}, \item{length}{description length}
}
\description{
Check coordinates in dat files.
}
\details{
Check if coordinates in all dat files in a directory are identical. Return values: "upstream", "downstream" and "length".
}
\examples{
    coord <- check.coord(datdir="/directory/containing/dat_files")
    upstream_value <- coord$upstream
    downstream_value <- coord$downstream
    length_value <- coord$lenght

}
\author{
Mark Boltengagen \email{m.boltengagen@gmail.com}
}
