% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.heatmaps.R
\name{combine.heatmaps}
\alias{combine.heatmaps}
\title{Combine heatmaps png into one image}
\usage{
combine.heatmaps(indir, outdir)
}
\arguments{
\item{indir}{Directory containing png heatmaps.}

\item{outdir}{Output directory.}
}
\description{
Combine heatmaps png into one image
}
