% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.avg.plot.R
\name{features.avg.plot}
\alias{features.avg.plot}
\title{All features averages plot}
\usage{
features.avg.plot(indir, outdir, fdir, chromsizes = "sc",
  lables = NULL, title = "All features averages plot",
  smooth = FALSE, log = FALSE, nlog = NULL, log.num = NULL,
  xlab = "Time", ylab = "Signal intensity",
  filename = "All_features_avg_plot", exclude_seq = "chrM",
  ylim = NULL)
}
\arguments{
\item{indir}{Input directory containing bedgraph files.}

\item{outdir}{Output directory.}

\item{fdir}{Directory containing different features in files.}

\item{chromsizes}{Text file containing sizes of all chromosomes in columns (chromosome ~ size).}

\item{title}{Main title of the plot.}

\item{smooth}{If lines between points are smoothed.}

\item{log}{Boolean, if log difference of signal is needed.}

\item{nlog}{Id of sample in files in 'indir' which is used for log difference (e.g. 1 - first sample) 'log' must be TRUE, log.num must be NULL.}

\item{log.num}{Numeric to calculate log difference (e.g. 21.5). 'log' must be TRUE, nlog must be NULL.}

\item{filename}{Name of the output plot file}

\item{exclude_seq}{Optional argument to exclude chromosomes.}
}
\description{
All features averages plot
}
