% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.box.plot.R
\name{features.box.plot}
\alias{features.box.plot}
\title{All features box plot}
\usage{
features.box.plot(indir, outdir, fdir, chromsizes,
  title = "Average log2 difference of H4K16Ac vs 1.5h sample",
  filename = "All_features_box_plot", notes = "",
  exclude_seq = "chrM")
}
\arguments{
\item{indir}{Input directory containing bedgraph files.}

\item{outdir}{Output directory.}

\item{fdir}{Directory containing different features in files.}

\item{chromsizes}{Text file containing sizes of all chromosomes in columns (chromosome ~ size).}

\item{title}{Main title of the plot}

\item{filename}{Name of the output plot file}

\item{notes}{Optional notes, which is displayed in the bottom fo the plot.}

\item{exclude_seq}{Optional argument to exclude chromosomes.}
}
\description{
All features box plot
}
