% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{find.norm.k}
\alias{find.norm.k}
\title{Calculate normalizing coefficients}
\usage{
find.norm.k(fbam, fbams, outdir, filename = "normfactors.txt")
}
\arguments{
\item{fbam}{Directory containing filtered bam files for peak calculaton.}

\item{fbams}{Directory containing filtered bam files of spiking using for normalizing.}

\item{outdir}{Output directory.}

\item{filename}{Name of output file.}
}
\description{
Calculate normalizing coefficients
}
