% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{genelength.from.sgd}
\alias{genelength.from.sgd}
\title{Get lengths of genes from SGS txt file}
\usage{
genelength.from.sgd(file, outdir, filename = "genelength", from = NULL,
  to = NULL)
}
\arguments{
\item{file}{SGS text file.}

\item{outdir}{Output directory.}

\item{filename}{Optional argument of file name.}

\item{from}{Shortest length to filter.}

\item{to}{Longest length to filter.}
}
\value{
None
}
\description{
Get lengths of genes from SGS txt file
}
\details{
Read sgd file, where name of genes (name), start of genes (txStart)
  and end of genes (txEnd) are specified. Write output text file with
  name~length columns.
}
