% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr.extra.R
\name{get.signal}
\alias{get.signal}
\title{Get signal}
\usage{
get.signal(gr, chr, start, end)
}
\arguments{
\item{gr}{Genomic Range containing metacolumn.}

\item{chr}{Chromosome (e.g. "chrI").}

\item{start}{Genomic coordinate to start get signal.}

\item{end}{Genomic coordinate to end get signal.}
}
\description{
Get signal from metacolumn of genomic ranges based on given chromosome, start and end genomic coordinates.
Return vector containing signal per nucleotide within a certain region.
}
\seealso{
\code{\link{coverage.plot}}
}
