% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr.extra.R
\name{gr.extend}
\alias{gr.extend}
\title{Extend Genomic Range depending on strand}
\usage{
gr.extend(x, upstream = 0, downstream = 0)
}
\arguments{
\item{upstream}{Number of nucleotides to extend upstream.}

\item{downstream}{Number of nucleotides to extend downstream.}
}
\value{
Genomic Range
}
\description{
Extend Genomic Range depending on strand
}
