% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr.extra.R
\name{gr.extend.reads}
\alias{gr.extend.reads}
\title{Extend GRange and trim}
\usage{
gr.extend.reads(x, fragmentLen = 75, trim = 0)
}
\value{
Genomic Ranges
}
\description{
Extend GRange to the certain length and trim till the trim length if necessary.
}
