% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.features.R
\name{meta.ars}
\alias{meta.ars}
\title{ARSs regions for metaplot}
\usage{
meta.ars <- function (ars_csv = "/home/suvar/Projects/Sources/Cs_genomic_features_v1/ARS.csv.gz",
    outdir = "/home/suvar/Projects/Sources/Cs_genomic_features_v1/meta/",
    win = 4000,
    chrom_sizes = "/home/suvar/Projects/Sources/sacCer3.chrom.sizes")
}
\arguments{
\item{ars_csv}{Text file containing ranges.}

\item{outdir}{Output directory.}

\item{win}{Windows around ARS's centers.}

\item{chrom_sizes}{File with sizes of chromosomes (.chrom.sizes).}
}
\value{
None
}
\description{
ARSs regions for metaplot
}
\details{
Create a csv.gz file containing saved Genomic Range of regions around ARSs.
}
\seealso{
\code{\link{meta.tss}}
}
