% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.data.R
\name{meta.data}
\alias{meta.data}
\title{Meta data generator}
\usage{
meta.data(meta.csv, bdg.data, outdir, upstream, downstream, chromsizes,
  threads = NULL)
}
\arguments{
\item{meta.csv}{csv file with meta features.}

\item{bdg.data}{Directory with bedgraph files of signal peaks (e.g. after MACS2 peak calling).}

\item{upstream}{Upstream nucleotides from feature.}

\item{downstream}{Downstrem nucleotides from feature.}

\item{chromsizes}{Text file containing sizes of all chromosomes in columns (chromosome ~ size).}

\item{threads}{Optional parameter of number of working processors. If not set, all-1 will be taken.}
}
\value{
TSS regions as CSV file dat.gz
}
\description{
Generate meta data for metaplots.
}
