% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.mean.R
\name{meta.mean}
\alias{meta.mean}
\title{Mean data file generator for meta plot.}
\usage{
meta.mean(datdir, outdir, file_name = "Meandata", split = c("no", "limits",
  "quantiles", "groups"), table = NULL, val.name = NULL, limits = NULL,
  ngr = 4)
}
\arguments{
\item{datdir}{Directory containing dat files generated by meta.dat.
function}

\item{outdir}{Output directory.}

\item{file_name}{Name of output file (optional).}

\item{split}{Argument showing how to split: 'no' - take all data (default);
'quantiles' - split by quantiles q1-q4; 'limits' - split by given vector of
values.}

\item{table}{Data table containing list of names in one column and a column
of sorted values. Must be set if split by either quantiles or limits.}

\item{limits}{A vector of values to split, e.g. c(0, 25, 50, 75, 100) will
split into for groups 0-25, 25-50, 50-75, 75-100, where 0 is a minimal
value and 100 is maximal.}
}
\description{
Mean data file generator for meta plot.
}
\seealso{
\code{\link{meta.data}}
}
\author{
Mark Boltengagen \email{m.boltengagen@gmail.com}
}
