% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.features.R
\name{meta.tss}
\alias{meta.tss}
\title{TSSs regions for metaplot}
\usage{
meta.tss(outdir, upstream, downstream, chrom_sizes, exclude_chrom = NULL)
}
\arguments{
\item{outdir}{Output directory.}

\item{upstream}{Number of nucleotides upstream from TSSs.}

\item{downstream}{Number of nucleotides downstream from TSSs (transcriptional start sites).}

\item{chrom_sizes}{File with sizes of chromosomes (.chrom.sizes).}
}
\value{
None
}
\description{
TSSs regions for metaplot
}
\details{
Create a csv.gz file containing saved Genomic Range of regions around TSSs (transcriptional start sites).
}
\seealso{
\code{\link{meta.ars}}
}
