% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{mxkmink}
\alias{mxkmink}
\title{Find normalizing coefficients based on Nucleosome Free Regions (NFR).
For data frames where rownames contain coordinates, and data frame contains data only.}
\usage{

  mxkmink(indir = c("/home/suvar/Projects/007_ChIPseq_HU-SR/output/tss_data/Et/means/all/Meandata-All_1.mean.dat.gz",
  "/home/suvar/Projects/007_ChIPseq_HU-SR/output/ars_dat/all/Et/HU_SR_Et_ars_meandata_All_1.mean.dat.gz"),
  snames = c("2h", "1.5h", "0'", "10'", "20'", "30'", "40'", "50'",
  "60'", "70'", "80'", "90'"), dirnames = c("et_tss", "et_ars"),
  weights = c(4547, 352))
}
\arguments{
\item{snames}{Sample names}

\item{dirnames}{Vector containing names of features (one feature in the input directory)}

\item{weights}{Vector containings number of every feature in a genome. Need for calculation of weighted means.}

\item{dir}{Directory containing vector of meandata for all features containing Nucleosome Free Regions.
e.g. TSS, ARS}
}
\value{
matrix of coefficients for each sample in each given feature
}
\description{
Find normalizing coefficients based on Nucleosome Free Regions (NFR).
For data frames where rownames contain coordinates, and data frame contains data only.
}
