% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxl.dotplot.R
\name{mxl.dotplot}
\alias{mxl.dotplot}
\title{Dot plot from list of matrices.

Take list of matrices and plot each matrix on separate figure. Legends are in name of matrices.}
\usage{
mxl.dotplot(mxl, outdir, filename, title, xlab = "Signal intensity",
  ylab = "Time series", ...)
}
\arguments{
\item{mxl}{List of matrices.}

\item{outdir}{Output directory.}

\item{filename}{Name of output file.}

\item{...}{Optional graphical parameters.}
}
\value{
None
}
\description{
Dot plot from list of matrices.

Take list of matrices and plot each matrix on separate figure. Legends are in name of matrices.
}
\author{
Mark Boltengagen \email{m.boltengagen@gmail.com}
}
