% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxl.dotplot.R
\name{mxl}
\alias{mxl}
\title{Create list of matrices from several experiments.

Create list of matrices, where length of list is number of files (splits) in directories.
For each matrix columns contains timepoints and rows contain experiments.}
\usage{
mxl(indir, expnames = NULL, timepoints = NULL, splitnames = NULL)
}
\arguments{
\item{indir}{Input directories. Each directory correspont to one experiment.}

\item{expnames}{Names of experiments. Names will be detected from names of directories if not specified.}

\item{timepoints}{Datasets in columns of each file. Names of datapoints will be detected from column names of the first file.}

\item{splitnames}{Names of files in directories. Splitnames will be detected from names of files in the first directory.}
}
\value{
List of matrices
}
\description{
Create list of matrices from several experiments.

Create list of matrices, where length of list is number of files (splits) in directories.
For each matrix columns contains timepoints and rows contain experiments.
}
\author{
Mark Boltengagen \email{m.boltengagen@gmail.com}
}
