% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{norm.mean.dat}
\alias{norm.mean.dat}
\title{Split files in a directory into groups}
\usage{
norm.mean.dat(norm, mdat, outdir)
}
\arguments{
\item{norm}{File containing normalizing coefficients (use find.norm.k function).}

\item{mdat}{Directory containing mean dat files which should be normalized.}

\item{outdir}{Output directory for normalized mean dat file.}
}
\description{
Split files in a directory into groups
}
