% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr.extra.R
\name{read.csv.gr}
\alias{read.csv.gr}
\title{Import csv to genomic ranges}
\usage{
read.csv.gr(file, chromsizes = "sc")
}
\arguments{
\item{chromsizes}{File containing chromosome sizes.
'sc' - Preset Saccharomyces serevisiae chromosome sizes.}

\item{CSV}{file containing tab separated columns: seqname, start, end, (optional: width), strand and optional extra 
columns containing metadata (e.g. genenemes, score, signal, cg). Read also csv files from .gz archives.
| seqname | start | end | strand | score |
| chrI    | 100   | 200 | +      | 1.5   |}
}
\description{
Read csv file and return genomic ranges.
}
\seealso{
\code{\link{saveGRlist}}
}
