% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.mean.R
\name{read.data.table}
\alias{read.data.table}
\title{Data table from text file}
\usage{
read.data.table(file, val.name)
}
\arguments{
\item{file}{txt file features in the first column ~ named corresponding values in columns}

\item{val.name}{Value name we would like to extract}
}
\value{
Data table with two columns (feature_name ~ corresponding_value)
}
\description{
Data table from text file
}
\details{
Read text file containig feature name in the first column and named
  columns of values. Function selects the column with names and the column with
  val.name, and return data.table sorted by val.name column (values).
}
