% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr.extra.R
\name{saveGR}
\alias{saveGR}
\title{Save Genomic Range as a csv file.}
\usage{
saveGR(gr, outdir = NULL, filename = "new_genomic_range")
}
\arguments{
\item{gr}{Genomic range}

\item{outdir}{Output directory}

\item{filename}{Name of output file}
}
\value{
None
}
\description{
Save Genomic Range as a csv file.
}
