% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr.extra.R
\name{saveGRlist}
\alias{saveGRlist}
\title{Save Genomic Ranges List as a set of csv files.}
\usage{
saveGRlist(grl, outdir = NULL)
}
\arguments{
\item{grl}{Named genomic ranges list}

\item{oudir}{Output directory.}
}
\value{
None
}
\description{
Save Genomic Ranges List as a set of csv files.
}
