% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr.extra.R
\name{seqlevels.from.chrsizes}
\alias{seqlevels.from.chrsizes}
\title{Get seqlevels from chromosome sizes txt file}
\usage{
seqlevels.from.chrsizes(chromsizes)
}
\arguments{
\item{chromsizes}{Tab separated text file containing sizes of all chromosomes in columns (chromosome ~ size).}
}
\value{
seql (named vector)
}
\description{
Transform chromosome sizes in a text file in genomic ranges seqlevels
}
