% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{setmink.dat}
\alias{setmink.dat}
\title{Normalize data in data frame to minimal value (Null value) by coefficient.}
\usage{
setmink.dat(df)
}
\arguments{
\item{df}{Data frame from mean data file.}
}
\value{
data frame (mean data)
}
\description{
Normalize data in data frame to minimal value (Null value) by coefficient.
}
