% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{setmink2.dat}
\alias{setmink2.dat}
\title{Normalize data in data frame to minimal value (Null value) by coefficient.
For data frames where rownames contain coordinates, and data frame contains data only.}
\usage{
setmink2.dat(df)
}
\arguments{
\item{df}{Data frame from mean data file.}
}
\value{
data frame (mean data)
}
\description{
Normalize data in data frame to minimal value (Null value) by coefficient.
For data frames where rownames contain coordinates, and data frame contains data only.
}
