% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.features.R
\name{split_intergenic}
\alias{split_intergenic}
\title{Split intergenic regions and save as csv genomic ranges:
    intergenic - all intergenics
    convergent - convergent intergenics
    devergent - devergent intergenics
    direct - direct intergenics (strand '+')
    reverse - reverse intergenics (strand '-')
    direct_reverse - combined direct and reverse intergenics}
\usage{
split_intergenic(outdir)
}
\arguments{
\item{outdir}{Output directory.}
}
\value{
None
}
\description{
Split intergenic regions and save as csv genomic ranges:
    intergenic - all intergenics
    convergent - convergent intergenics
    devergent - devergent intergenics
    direct - direct intergenics (strand '+')
    reverse - reverse intergenics (strand '-')
    direct_reverse - combined direct and reverse intergenics
}
\details{
Create a csv.gz file containing saved Genomic Range of split intergenic regions.
}
