% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.mean.R
\name{write.mean.dat}
\alias{write.mean.dat}
\title{Write mean dat file.}
\usage{
write.mean.dat(datdir, selected_names = NULL, outdir, file_name)
}
\arguments{
\item{datdir}{Directory containing dat files generated by meta.dat.
function}

\item{selected_names}{Vector of names wich will be used to calculate mean values.}

\item{outdir}{Output directory.}

\item{file_name}{Name of output file.}
}
\description{
Write mean dat file.
}
\author{
Mark Boltengagen \email{m.boltengagen@gmail.com}
}
