% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bamcov.plot.R, R/bamcov.plot2.R
\name{bamcov.plot}
\alias{bamcov.plot}
\title{Bam Coverage Statter Plot}
\usage{

  bamcov.plot(indir1 = "/home/suvar/Projects/007_ChIPseq_HU-SR/output/bam_cov/Et/IP",
  indir2 = "/home/suvar/Projects/007_ChIPseq_HU-SR/output/bam_cov/IAA/IP",
  outdir = "/home/suvar/Desktop/", lables = c("HU-SR 2h", "HU-SR 0'",
  "HU-SR 10'", "HU-SR 20'", "HU-SR 30'", "HU-SR 40'", "HU-SR 50'",
  "HU-SR 60'", "HU-SR 70'", "HU-SR 80'", "HU-SR 90'"),
  filename = "BamCoverage_plot.png", log = FALSE)


  bamcov.plot(indir1 = "/home/suvar/Projects/007_ChIPseq_HU-SR/output/bam_cov/Et/IP",
  indir2 = "/home/suvar/Projects/007_ChIPseq_HU-SR/output/bam_cov/IAA/IP",
  outdir = "/home/suvar/Desktop/", lables = c("HU-SR 2h", "HU-SR 0'",
  "HU-SR 10'", "HU-SR 20'", "HU-SR 30'", "HU-SR 40'", "HU-SR 50'",
  "HU-SR 60'", "HU-SR 70'", "HU-SR 80'", "HU-SR 90'"),
  filename = "BamCoverage_plot.png", log = FALSE)
}
\arguments{
\item{outdir}{Output directory}

\item{indir}{Directory containing bam coverage binned bedgraph files}

\item{indir}{Directory containing bam coverage binned bedgraph files}

\item{outdir}{Output directory}
}
\description{
Bam Coverage Statter Plot

Bam Coverage Statter Plot
}
