% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.avg.plot.R
\name{features.avg.plot}
\alias{features.avg.plot}
\title{All features averages plot}
\usage{
features.avg.plot(indir, outdir, fdir, chromsizes = "sc",
  lables = NULL, title = "All features averages plot", log = FALSE,
  nlog = NULL, xlab = "Time", ylab = "Signal intensity",
  filename = "All_features_avg_plot", exclude_seq = "chrM",
  ylim = NULL)
}
\arguments{
\item{indir}{Input directory containing bedgraph files.}

\item{outdir}{Output directory.}

\item{fdir}{Directory containing different features in files.}

\item{chromsizes}{Text file containing sizes of all chromosomes in columns (chromosome ~ size).}

\item{title}{Main title of the plot}

\item{filename}{Name of the output plot file}

\item{exclude_seq}{Optional argument to exclude chromosomes.}

\item{notes}{Optional notes, which is displayed in the bottom of the plot.}
}
\description{
All features averages plot
}
