% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.win.avg.R
\name{heatmap.win.avg}
\alias{heatmap.win.avg}
\title{Heatmap of binned averages from bedgraph files}
\usage{
heatmap.win.avg(indir, outdir, chromosome = c(1, 2, 3, 4, 5, 6, 7, 8, 9,
  10, 11, 12, 13, 14, 15, 16), norm = NULL, fname = "HeatMap",
  log = TRUE, nlog = NULL, width = NULL, step = NULL,
  comb = TRUE, threads = 2)
}
\arguments{
\item{indir}{Directory containing bedgraph files with peaks}

\item{outdir}{Output directory}

\item{chromosome}{Chromosome to draw}

\item{norm}{Vector of normalization coeficients}

\item{fname}{File name}

\item{log}{Log difference (to common average)}

\item{nlog}{Given number to compair (e.g. global mean). It is not calculated if define.}

\item{comb}{Create pdf file with combined heatmaps for all chromosomes (all png in the output directory)}

\item{threads}{How many logical processors are available?}
}
\description{
Heatmap of binned averages from bedgraph files
}
\details{
Heatmap plot of binned averages from bedgraph files
}
