% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.plot.R
\name{meta.plot}
\alias{meta.plot}
\title{TSS & ARS regions for metaplot

These functions allows to create txt file with a list of TSS or ARS with a given size.}
\usage{
meta.plot(indir = "/home/suvar/Projects/001_ChIPseq_HU/output/tss_dat",
  outdir, col.file, file_name = "Meta_plot", legend_names, lines_to_draw,
  main = "Meta Plot", xlab = "Coordinates", ylab = "Signal",
  xlim = NULL, ylim = NULL, las = 1, cex.main = 1, cex.lab = 1,
  cex.axis = 1, cex = 1)
}
\arguments{
\item{col.file}{File containing list of colors in the first column}

\item{legend_names}{A vector with legend names.}

\item{main}{Main header of a plot.}

\item{xlab}{downstrem nucleotides from feature.}

\item{dir}{Input directory with DAT files (dat.gz).}

\item{...}{Optional graphics arguments}
}
\value{
None
}
\description{
TSS & ARS regions for metaplot

These functions allows to create txt file with a list of TSS or ARS with a given size.
}
\seealso{
\code{\link{meta.data}}, \code{\link{meandata}}
}
