% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.plot.R
\name{meta.plot}
\alias{meta.plot}
\title{TSS & ARS regions for metaplot

These functions allows to create txt file with a list of TSS or ARS with a
given size.}
\usage{
meta.plot(indir, outdir, col.file, filename = "Meta_plot", curnames = NULL,
  ncur = NULL, labels = NULL, log = FALSE, main = "Meta Plot",
  xlab = "Distance from feature, bp", ylab = "Signal", xlim = NULL,
  ylim = NULL, legendpos = c("bottomright", "bottom", "bottomleft", "left",
  "topleft", "top", "topright", "right", "center"), las = 1, cex.lab = 1,
  cex.axis = 1)
}
\arguments{
\item{indir}{Input directory with means DAT files (mean.dat.gz) generated by
\code{\link{meandata}} funcion.}

\item{outdir}{Output directory.}

\item{col.file}{Text file containing list of colors in the first column.}

\item{filename}{Name of output file.}

\item{curnames}{A vector containing desiered legend names. The order of names
must correspond to the data in the mean.dat files.}

\item{ncur}{Numeric vector of lines to draw (e.g  c(2,4,6)). The first
element in 'ncur' is a denominator if 'log' argument is TRUE.}

\item{labels}{Vector of names for multiple plot.}

\item{log}{Log2 of difference. The denominator is the first element in 'ncur'
argument.}

\item{main}{Main header of the plot.}

\item{xlab}{X-axis lable.}

\item{ylab}{Y-axis lable.}

\item{xlim}{The x limits (x1, x2) of the plot.}

\item{ylim}{The y limins (y1, y2) of the plot.}

\item{legendpos}{Position of the legend. Possible values: "bottomright",
"bottom", "bottomleft", "left", "topleft", "top", "topright", "right",
"center".}
}
\value{
None
}
\description{
TSS & ARS regions for metaplot

These functions allows to create txt file with a list of TSS or ARS with a
given size.
}
\seealso{
\code{\link{meta.data}}, \code{\link{meandata}}
}
\author{
Mark Boltengagen \email{m.boltengagen@gmail.com}
}
